requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0